% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk.R
\name{Singularity}
\alias{Singularity}
\title{R6 Class representing the Singularity SDK}
\description{
R6 Class representing the Singularity SDK

R6 Class representing the Singularity SDK
}
\details{
Allows configuration of the interactions between this client and the
Singularity Oauth 2.0 server

the custom message handler created to allow redirection from shiny
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{the}}{client id generated by the singularity server that represents this client}

\item{\code{client_secret}}{the client secret generated by the singularity server to allow this client to access the API}

\item{\code{client_key}}{the client key generated by the singularity server to allow this client to access the API}

\item{\code{message_handler}}{the name of the message handler to create for shiny js code}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{the}}{client id generated by the singularity server that represents this client}

\item{\code{client_secret}}{the client secret generated by the singularity server to allow this client to access the API}

\item{\code{client_key}}{the client key generated by the singularity server to allow this client to access the API}

\item{\code{message_handler}}{the name of the message handler to create for shiny js code}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Singularity$new()}}
\item \href{#method-print}{\code{Singularity$print()}}
\item \href{#method-shiny_tags}{\code{Singularity$shiny_tags()}}
\item \href{#method-shiny_redirect}{\code{Singularity$shiny_redirect()}}
\item \href{#method-clone}{\code{Singularity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the SDK
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Singularity$new(
  client_id,
  client_secret,
  client_key,
  config_file = ".local.config.json",
  config = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client_id}}{the client identifier as generated by Singularity}

\item{\code{client_secret}}{the client secret as generated by Singularity}

\item{\code{client_key}}{the client key as generated by Singularity}

\item{\code{config_file}}{optional configuration file to override the default configuration from}

\item{\code{config}}{optional config list to override the default configuration with}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints a representation of this object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Singularity$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{print parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-shiny_tags"></a>}}
\if{latex}{\out{\hypertarget{method-shiny_tags}{}}}
\subsection{Method \code{shiny_tags()}}{
Adds a script tag with the redirection handler
This is specifically for shiny
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Singularity$shiny_tags(tags)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tags}}{the shiny tags object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-shiny_redirect"></a>}}
\if{latex}{\out{\hypertarget{method-shiny_redirect}{}}}
\subsection{Method \code{shiny_redirect()}}{
Calls the shiny redirect handler
This is specifically for shiny
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Singularity$shiny_redirect(session, redirect_uri)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{session}}{the session object from shiny}

\item{\code{redirect_uri}}{the uri to redirect the user to in
order to force the trusted login flow}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Singularity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
